began fixing broken merge

+++ src/framebuffer.h
@@ -1,17 +1,18 @@
 #pragma once
+#include <stddef.h>
 
 namespace texture {
 
 	void initFramebuffers();
 
 	void initGBuffer();
-	extern size_t gBuffer;
-	extern size_t gPosTexture;
-	extern size_t gNormalTexture;
-	extern size_t gAmbientTexture;
-	extern size_t gDiffuseTexture;
-	extern size_t gSpecularTexture;
-	extern size_t gDepthRenderbuffer;
+	extern unsigned int gBuffer;
+	extern unsigned int gPosTexture;
+	extern unsigned int gNormalTexture;
+	extern unsigned int gAmbientTexture;
+	extern unsigned int gDiffuseTexture;
+	extern unsigned int gSpecularTexture;
+	extern unsigned int gDepthRenderbuffer;
 
 	void initGUIFBO();
 	extern unsigned int guiFBO;
@@ -19,3 +20,4 @@ namespace texture {
 	extern unsigned int fontColorTexture;
 	extern unsigned int guiDepthRenderbuffer;
 }
+
